Move types into a types.ts file
